feat(web,shared,docs): split the companion settings page into its own top-level route
The companion was the tallest page in settings: three cards, one
scroll, roughly 770 lines. I moved it out of Settings into its own
top-level sidebar group (Assistant) and split it into three routes,
each gating itself in its own +page.server.ts rather than inheriting a
gate from a layout.

- /companion (persona - handle, display name, headline, about, with
  the provenance badge) is the landing page, since that is what I
  correct most and it is now visible without scrolling.
- /companion/voice (the derived voice and its samples).
- /companion/work (GitHub repos and the GitHub App install panel -
  read as operator context by the prompt, not a project's own data).

/settings/companion stays a redirect (307) rather than 404ing, since
it was deep-linked from the GitHub App install round trip. That round
trip's BACK constant now points straight at /companion/work instead of
bouncing through the old route.

Nothing about what the prompt reads changed: loadCompanionContext and
buildSuggestionPrompt are untouched.

Recorded as D35 in docs/design/DECISIONS.md, refining D15. Updated
docs/permissions.md's route/role table and its settings-rail history,
and docs/design/linkedin-assistant-brief.md's route reference.

web/tests/companion-settings.test.ts now covers the three split
loaders/actions instead of one; the retired redirect and the GitHub
setup round trip's new destination are covered by
settings-general-redirect.test.ts and github-install-flow.test.ts.
